move UI utils functions in a new file
