+++ src/components/TaskItem.tsx
@@ -8,6 +8,7 @@ import { hookMarkdownLinkMouseEventHandlers, markdownButtonHoverPreviewEvent } f
 import { Component } from 'obsidian';
 import { EditButtonMode } from 'src/interfaces/GlobalSettings';
 import { MarkdownUIRenderer } from 'src/services/MarkdownUIRenderer';
+import { hexToRgba } from 'src/utils/UIHelpers';
 import { priorityEmojis } from '../interfaces/TaskItemProps';
 import { t } from 'src/utils/lang/helper';
 
@@ -166,23 +167,6 @@ const TaskItem: React.FC<TaskProps> = ({ app, plugin, taskKey, task, columnIndex
 			event.ctrlKey = false;
 		}
 	}
-
-	// Helper function to convert a hex color to an RGBA string with the specified opacity
-	const hexToRgba = (hex: string, opacity: number): string => {
-		let r = 0, g = 0, b = 0;
-
-		if (hex.length === 4) {
-			r = parseInt(hex[1] + hex[1], 16);
-			g = parseInt(hex[2] + hex[2], 16);
-			b = parseInt(hex[3] + hex[3], 16);
-		} else if (hex.length === 7 || hex.length === 9) {
-			r = parseInt(hex[1] + hex[2], 16);
-			g = parseInt(hex[3] + hex[4], 16);
-			b = parseInt(hex[5] + hex[6], 16);
-		}
-
-		return `rgba(${r},${g},${b},${opacity})`;
-	};
 
 	// Default color when tag isn't found in the settings
 	const defaultTagColor = 'var(--tag-color)';
